test(voice): cover exclusive WAV reopen
