+++ pinvou3-app/src-tauri/src/features/voice/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod microphone_permission;
 mod platform;
+mod temp_wav;
 mod transcript;
 pub(crate) mod voice_asr;
 
@@ -7,5 +8,6 @@ pub(crate) use platform::{
     asr_dependency_packages, asr_missing_message, asr_tool_exists, asr_tool_path,
     engine_binary_name, native_recognition_source, recognize_native,
 };
+pub(crate) use temp_wav::VoiceTempWav;
 pub(crate) use transcript::parse_asr_transcript;
 pub(crate) use voice_asr::set_bundled_engine_dir;
